Better error checking for pdf creation

+++ server.js
@@ -2,24 +2,18 @@ const express = require('express')
 const app = express()
 const bodyParser = require('body-parser')
 const http = require('http')
-const { exec } = require('child_process')
+const { exec, spawnSync } = require('child_process')
 const fs = require('fs')
-const crypto = require('crypto')
 const { translate_and_save: translate_z3 } = require('./src/translator/z3')
 const { translate_to_file: translate_latex } = require('./src/translator/latex')
 const { parse } = require('./src/parser')
+const { random_file_name } = require('./src/utils')
 
 // Configuration
 app.use(bodyParser.urlencoded({ extended: true }))
 app.use(bodyParser.json())
 
 app.use(express.static(__dirname + '/public'))
-
-const random_file_name = () => {
-  const current_date = (new Date()).valueOf().toString()
-  const random = Math.random().toString()
-  return crypto.createHash('sha1').update(current_date + random).digest('hex').toString()
-}
 
 app.post('/z3', (req, res) => {
   const { steps, atoms, constants, relations } = req.body
@@ -40,14 +34,25 @@ app.post('/pdf', (req, res) => {
   const contents = translate_latex(givens, steps)
   const filename = random_file_name()
   fs.writeFileSync(filename, contents)
-  exec('./pdflatex ' + filename, (err, stdout) => {
-    const pdf = fs.readFileSync(filename + '.pdf')
-    fs.unlinkSync(filename)
-    fs.unlinkSync(filename + '.pdf')
+  spawnSync('pdftex', [filename], {stdio: 'ignore'})
+  spawnSync('pdflatex', [filename], {stdio: 'ignore'})
+  fs.unlinkSync(filename)
+  fs.unlinkSync(filename + '.log')
+  if (fs.existsSync(filename + '.aux')) {
     fs.unlinkSync(filename + '.aux')
-    fs.unlinkSync(filename + '.log')
+  }
+  if (fs.existsSync('texput.log')) {
+    fs.unlinkSync('texput.log')
+  }
+  if (fs.existsSync(filename + '.pdf')) {
+    const pdf = fs.readFileSync(filename + '.pdf')
+    fs.unlinkSync(filename + '.pdf')
+    
     res.send(pdf)
-  })
+  }
+  else {
+    res.status(404).end()
+  }
 })
 
 // Start server
